refactor(fe): rename admin route to setting
