+++ frontend/src/Router.tsx
@@ -16,11 +16,11 @@ export default class Router extends Component {
         <Route path="/notice/:id" element={<Screen.NoticeDetailScreen />} />
         <Route path="/notice/create" element={<Screen.NoticeCreateScreen />} />
         <Route path="/register" element={<Screen.RegisterScreen />} />
-        <Route path="/admin" element={<Screen.SettingScreen />} />
-        <Route path="/admin/schedule" element={<Screen.ScheduleSettingScreen />} />
-        <Route path="/admin/fg" element={<Screen.FgSettingScreen />} />
-        <Route path="/admin/freshman" element={<Screen.FreshmanSettingScreen />} />
-        <Route path="/admin/lcdate" element={<Screen.LcDateSettingScreen />} />
+        <Route path="/setting" element={<Screen.SettingScreen />} />
+        <Route path="/setting/schedule" element={<Screen.ScheduleSettingScreen />} />
+        <Route path="/setting/fg" element={<Screen.FgSettingScreen />} />
+        <Route path="/setting/freshman" element={<Screen.FreshmanSettingScreen />} />
+        <Route path="/setting/lcdate" element={<Screen.LcDateSettingScreen />} />
       </Routes>
     )
   }
