+++ frontend/src/screen/setting/component/loading.tsx
