+++ frontend/src/screen/setting/freshmanSetting.tsx
